+++ colorkeys/createjson.py
@@ -1,12 +1,14 @@
 #!/usr/bin/env python3
 
-import datetime
 import hashlib
 import json
 import logging
 import pkg_resources
 import re
 import sys
+from datetime import datetime
+from datetime import timezone
+from pathlib import Path
 from urllib import request
 
 from engcommon import command
@@ -42,8 +44,8 @@ def compile(palette, **kwargs):
             "hist_centroids": h.hist_centroids,
         })
     obj = {
-        "filename": palette.imgsrc,
-        "sha1sum": get_sha1(palette.imgsrc),
+        "filename": Path(palette.imgsrc).name,
+        "filehash": get_filehash(palette.imgsrc),
         "shape": palette.img.shape,
         "timestamp": get_timestamp(),
         "version": get_version(pkg_name),
@@ -53,6 +55,7 @@ def compile(palette, **kwargs):
     if my_aws:
         obj["cpu"] = my_aws.task_desc["cpu"]
         obj["memory"] = my_aws.task_desc["memory"]
+        obj["task_hash"] = my_aws.task_hash
     return obj
 
 
@@ -68,11 +71,7 @@ def get_version(pkg_name):
 
 def get_timestamp():
     """Get current timestamp."""
-    dt = datetime.datetime.now()
-    timestamp = (
-        f"{dt.year}.{dt.month:02d}.{dt.day:02d}-"
-        f"{dt.hour:02d}{dt.minute:02d}{dt.second:02d}"
-    )
+    timestamp = datetime.now(timezone.utc).astimezone().isoformat()
     return timestamp
 
 
@@ -97,22 +96,22 @@ def get_githash(pkg_name):
     return githash
 
 
-def get_sha1(filename):
-    """Get SHA1 of file.
+def get_filehash(filename):
+    """Get hash of file.
 
     Args:
         filename (str): File name.
 
     Returns:
-        sha1sum (str): SHA1 checksum of file.
+        filehash (str): hash digest of file.
     """
     BLOCK_SIZE = 65536
-    sha1sum = hashlib.sha1()
+    filehash = hashlib.blake2b(digest_size=8)
     prefixes = ("http://", "https://")
     with request.urlopen(filename) \
             if filename.startswith(prefixes) else open(filename, "rb") as f:
         fb = f.read(BLOCK_SIZE)
         while len(fb) > 0:
-            sha1sum.update(fb)
+            filehash.update(fb)
             fb = f.read(BLOCK_SIZE)
-    return sha1sum.hexdigest()
+    return filehash.hexdigest()